issue:master description: - Finalizando desafio

+++ weather_app/lib/app/models/weather_response.dart
@@ -160,10 +160,10 @@ class Main {
   });
 
   Main.fromJson(Map<String, dynamic> json) {
-    temp = json['temp'];
-    feelsLike = json['feels_like'];
-    tempMin = json['temp_min'];
-    tempMax = json['temp_max'];
+    temp = json['temp'].toDouble();
+    feelsLike = json['feels_like'].toDouble();
+    tempMin = json['temp_min'].toDouble();
+    tempMax = json['temp_max'].toDouble();
     pressure = json['pressure'];
     humidity = json['humidity'];
     seaLevel = json['sea_level'];
